Update README.md
Linear Regression with Python Scikit Learn
In this section, we will see how the Python Scikit-Learn library for machine learning can be used to implement regression functions. We will start with simple linear regression involving two variables.

Linear Regression with Python Scikit Learn
In this section, we will see how the Python Scikit-Learn library for machine learning can be used to implement regression functions. We will start with simple linear regression involving two variables.

Problem statement
In this regression task, we will predict the percentage of marks that a student is expected to score based on the number of hours they studied. This is a simple linear regression task as it involves just two variables.

To predict:
What will be the predicted score if a student studies for 9.25 hrs/ day?